add support for Ubuntu Focal, set deps as variable
The current list of packages is hard-coded and fails when deploying to
Ubuntu Focal which is Python 3 only and has some renamed packages (such
as gunicorn3 -> gunicorn). The Ansible PPA is also not required on
Focal, nor is it available and so actually fails.

This patch skips adding the Ansible PPA on Focal or later release and
also turns the list of dependencies into a variable, stored in a distro
specific vars file (vars/ubuntu.yml). This makes it easier to put logic
into the list as dependencies change in the future and also allows users
to override the package list if needed, without modifying the code.

It has an added benefit of making it easier to support other distros in
the future, if so desired.